Add -i option to produce ramp-up results
Adds the -i option (--interval) to supply a time interval to
repeatedly report results (requests/sec and bytes) after each
interval until the full time duration is up. The main thread
is used to collect each thread's intermediate results since
the last reported interval result and to report this interval's
results on one line.

Not suitable for 32-bit platforms as the values are not read
atomically. In practice, it works pretty well but could be made
more thread safe.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>